Add files via upload
Added push button

but I am not really sure yet that this will work. Need to hook it up to the hardware to see how it does. 
There will probably be a problem with debounce and the amount of time that a number shows on the dice.

Also see Single_Random_Number_on_Button_Push.ino for an incredibly simple example of debounce that probably won't work here.